requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==2.1.3
requests==2.32.3
gspread==6.1.4
google-auth==2.35.0